+++ src/app/(user)/components/reservation/RoomInformation.tsx
