+++ src/app/(user)/recently-viewed/components/RoomCard.tsx
