+++ src/app/(user)/trips/components/Help.tsx
