+++ src/app/(user)/trips/components/NoReservation.tsx
