+++ src/app/(user)/trips/components/ReservationCard.tsx
